added keep_tracks_seperate, preparing for release

+++ scripts/fastVideo.py
@@ -23,7 +23,7 @@ from shutil import rmtree
 from time import time
 
 def fastVideo(videoFile, outFile, silentThreshold, frameMargin, SAMPLE_RATE,
-    AUD_BITRATE, VERBOSE, cutByThisTrack):
+    AUD_BITRATE, VERBOSE, cutByThisTrack, keepTracksSep):
 
     print('Running from fastVideo.py')
 
@@ -129,23 +129,35 @@ def fastVideo(videoFile, outFile, silentThreshold, frameMargin, SAMPLE_RATE,
 
     # Now mix new audio(s) and the new video.
 
-    if(tracks > 1):
-        cmd = ['ffmpeg']
+    if(keepTracksSep):
+        cmd = ['ffmpeg', '-y']
         for i in range(tracks):
             cmd.extend(['-i', f'{TEMP}/new{i}.wav'])
-        cmd.extend(['-filter_complex', f'amerge=inputs={tracks}', '-ac', '2',
-            f'{TEMP}/newAudioFile.wav'])
+        cmd.extend(['-i', f'{TEMP}/spedup.mp4']) # add input video
+        for i in range(tracks):
+            cmd.extend(['-map', f'{i}:a:0'])
+        cmd.extend(['-map', f'{tracks}:v:0','-c:v', 'copy', '-movflags', '+faststart',
+            outFile])
+        if(not VERBOSE):
+            cmd.extend(['-nostats', '-loglevel', '0'])
+    else:
+        if(tracks > 1):
+            cmd = ['ffmpeg']
+            for i in range(tracks):
+                cmd.extend(['-i', f'{TEMP}/new{i}.wav'])
+            cmd.extend(['-filter_complex', f'amerge=inputs={tracks}', '-ac', '2',
+                f'{TEMP}/newAudioFile.wav'])
+            if(not VERBOSE):
+                cmd.extend(['-nostats', '-loglevel', '0'])
+            subprocess.call(cmd)
+        else:
+            os.rename(f'{TEMP}/new0.wav', f'{TEMP}/newAudioFile.wav')
+
+        cmd = ['ffmpeg', '-y', '-i', f'{TEMP}/newAudioFile.wav', '-i',
+            f'{TEMP}/spedup.mp4', '-c:v', 'copy', '-movflags', '+faststart',
+            outFile]
         if(not VERBOSE):
             cmd.extend(['-nostats', '-loglevel', '0'])
         subprocess.call(cmd)
-    else:
-        os.rename(f'{TEMP}/new0.wav', f'{TEMP}/newAudioFile.wav')
-
-    cmd = ['ffmpeg', '-y', '-i', f'{TEMP}/newAudioFile.wav', '-i',
-        f'{TEMP}/spedup.mp4', '-c:v', 'copy', '-movflags', '+faststart',
-        outFile]
-    if(not VERBOSE):
-        cmd.extend(['-nostats', '-loglevel', '0'])
-    subprocess.call(cmd)
 
     return outFile
